+++ test/query_parser_test.dart
@@ -47,4 +47,30 @@ void main() {
       expect(f, equals(filter));
     });
   });
+
+  // issue #65
+  test('Special characters in filter', () {
+    String dn = r'téstèr';
+    var q = '(roleOccupant=CN=$dn)';
+
+    var f = parseQuery(q);
+
+    expect(f, isA<Filter>());
+    expect(f.filterType, equals(Filter.TYPE_EQUALITY));
+    expect(f.attributeName, equals('roleOccupant'));
+    var x = ASN1OctetString(r'CN=t\C3\A9st\C3\A8r');
+
+    expect(f.assertionValue, equals(x));
+  });
+
+  test('escaping characters', () {
+    var s = r'téstèr';
+    var expected = r't\C3\A9st\C3\A8r';
+    var escaped = escapeNonAscii(s);
+    expect(escaped, equals(expected));
+
+    var t = r'(roleOccupant=cn=fred\5c, _smith,ou=users,dc=example,dc=com)';
+    expect(t, equals(escapeNonAscii(t)),
+        reason: 'escaped chars should not be escaped again');
+  });
 }