+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Dartdap Change Log
+## 0.8.0
+- Added a utility to escape charcters in LDAP strings
+- Updated ldap.query to support escaping of special characters
+- Added tests for #65
+
 ## 0.7.6
 
 - Update asn1lib.